attachments

+++ src/infra/database/prisma/repositories/prisma-question-attachments-repository.ts
@@ -2,6 +2,7 @@ import { QuestionAttachmentsRepository } from '@/domain/forum/application/reposi
 import { Injectable } from '@nestjs/common'
 import { PrismaService } from '../prisma.service'
 import { PrismaQuestionAttachmentMapper } from '../mappers/prisma-question-attachment-mapper'
+import { QuestionAttachment } from '@/domain/forum/enterprise/entities/question-attachment'
 
 @Injectable()
 export class PrismaQuestionAttachmentsRepository
@@ -9,17 +10,51 @@ export class PrismaQuestionAttachmentsRepository
 {
   constructor(private prisma: PrismaService) {}
 
-  async findManyByQuestionId(questionId: string) {
+  async findManyByQuestionId(
+    questionId: string,
+  ): Promise<QuestionAttachment[]> {
     const questionAttachments = await this.prisma.attachment.findMany({
-      where: { questionId },
+      where: {
+        questionId,
+      },
     })
 
     return questionAttachments.map(PrismaQuestionAttachmentMapper.toDomain)
   }
 
-  async deleteManyByQuestionId(questionId: string) {
+  async createMany(attachments: QuestionAttachment[]): Promise<void> {
+    if (attachments.length === 0) {
+      return
+    }
+
+    const data = PrismaQuestionAttachmentMapper.toPrismaUpdateMany(attachments)
+
+    await this.prisma.attachment.updateMany(data)
+  }
+
+  async deleteMany(attachments: QuestionAttachment[]): Promise<void> {
+    if (attachments.length === 0) {
+      return
+    }
+
+    const attachmentIds = attachments.map((attachment) => {
+      return attachment.id.toString()
+    })
+
     await this.prisma.attachment.deleteMany({
-      where: { questionId },
+      where: {
+        id: {
+          in: attachmentIds,
+        },
+      },
+    })
+  }
+
+  async deleteManyByQuestionId(questionId: string): Promise<void> {
+    await this.prisma.attachment.deleteMany({
+      where: {
+        questionId,
+      },
     })
   }
 }
